feat(#11): add last element to the planet1
