+++ Assets/Scripts/ActionManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4380eb4a89abeb248a5baa8a9208b066
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
