+++ Assets/Scripts/DialogueManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ce92e30d90290804eba6a3663db679f2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
