--- Assets/Scripts/Interactive Menu/SpashipMovement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 130558dd3cae3a3488cc618ee7f4edee
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
