--- Assets/Scripts/Planets/Action.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d595705041c51c74bbf24c24d2bc396d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
