--- Assets/Scripts/Planets/DetectionManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7f7fca587fe4bec4dbc2a72f489f660f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
