--- Assets/Scripts/Planets/Dialogue.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 55498ff63bf208947946bc6148feb2cd
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
