--- Assets/Scripts/Planets/PrincessMovement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4825ada93376d664bba897a2169f60ff
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
